# internal/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Go Demo CRUD API
  version: 1.0.0
  description: Simple CRUD API for products using Oracle (godror)
servers:
  - url: http://localhost:8080
paths:
  /health:
    get:
      tags: [Health]
      summary: Health check
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
  /products:
    get:
      tags: [Products]
      summary: List products
      responses:
        '200':
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
        '500':
          description: Server error
    post:
      tags: [Products]
      summary: Create product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductCreate'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          description: Invalid input
        '500':
          description: Server error
  /products/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer
          format: int64
        description: Product ID
    get:
      tags: [Products]
      summary: Get product by ID
      responses:
        '200':
          description: Product
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          description: Not found
        '500':
          description: Server error
    put:
      tags: [Products]
      summary: Update product by ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductUpdate'
      responses:
        '200':
          description: Updated product
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          description: Invalid input
        '404':
          description: Not found
        '500':
          description: Server error
    delete:
      tags: [Products]
      summary: Delete product by ID
      responses:
        '204':
          description: Deleted
        '404':
          description: Not found
        '500':
          description: Server error
components:
  schemas:
    Product:
      type: object
      properties:
        product_id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: Sample product
        attributes_varchar:
          type: string
          nullable: true
          example: size=L;color=blue
        attributes_clob:
          type: string
          nullable: true
          example: "{\"long\": \"json or text\"}"
        attributes_blob:
          type: string
          format: byte
          description: Base64-encoded binary
      required: [product_id, name]
    ProductCreate:
      type: object
      properties:
        name:
          type: string
        attributes_varchar:
          type: string
          nullable: true
        attributes_clob:
          type: string
          nullable: true
        attributes_blob:
          type: string
          format: byte
      required: [name]
    ProductUpdate:
      allOf:
        - $ref: '#/components/schemas/ProductCreate'